+++ src/main/web-portal/src/App.test.js
@@ -1,6 +1,8 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import App from './App';
-import {act, ReactTestUtils} from "react-dom/test-utils";
+import {act} from 'react-dom/test-utils';
+import {Simulate} from "react-dom/test-utils";
+
 // We use ReactTestUtils to simulate user input!
 
 function wait( milliseconds ) {
@@ -30,7 +32,7 @@ test('Test login form', async () => {
   // First username:
   // Let's type in the username so that it also triggers the onChange event
   await act(async () => {
-    await ReactTestUtils.Simulate.change(username, {target: {value: 'a'}});
+    await Simulate.change(username, {target: {value: 'a'}});
   });
 
   // We are doing MVVM, so let's wait for the view-model to update the feedback text
