+++ src/main/web-portal/src/mvvm/Constants.ts
